refactor: encapsulate GAME_WEB divergence
Consolidates the scattered #ifdef GAME_WEB sites (38 -> 31, with the
remainder being deliberate single-definition per-platform constants or
genuine flow divergences):

- New src/web_platform.{h,cc}: the only translation unit including
  <emscripten.h>. Hosts the IndexedDB sync machinery (moved from
  platform.{h,cc}) with inline desktop no-ops, and RunBrowserMainLoop /
  CancelBrowserMainLoop wrappers so game.cc no longer touches the
  Emscripten API directly.
- Networking is now selected by the build system instead of a
  whole-file #ifdef split: src/network.cc is native-only,
  src/network_web.cc is the browser stub, and CMake picks via
  ENGINE_PLATFORM_SRCS (same pattern as VENDORED_NATIVE_SRCS).
- sdl_init.cc: the duplicated per-platform GL-attribute CHECK block
  collapses to a GlContextConfig constant with one ifdef and a single
  set of CHECKs.
- save.cc: the journal/synchronous pragmas hoist to per-platform
  constants; Open() has one unconditional code path. Flush() keeps its
  ifdef (IDB sync vs WAL checkpoint genuinely differ).
- renderer.cc: GL_LINE_SMOOTH merges into the adjacent GL_MULTISAMPLE
  guard (order-independent glEnable calls).

No behavior change on either platform; verified via desktop suite +
flappybird run, web build, and headless-Firefox boot (renders, zero
errors, IndexedDB persistence across two sessions).
